Move exceptions into separate folder
