+++ src/DependencyGraph/IDependencyExecutionEngine_1.cs
@@ -9,6 +9,7 @@ using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Threading;
 using System.Threading.Tasks;
+using LanceC.DependencyGraph.Exceptions;
 
 namespace LanceC.DependencyGraph
 {
